+++ vbyte.h
@@ -20,8 +20,8 @@
 
 #include <stdint.h>
 
-int vbyte_read(const uint8_t *p, uint64_t *out);
-int vbyte_store(uint8_t *p, uint64_t value);
+int vbyte_read(const uint8_t *p, uint32_t *out);
+int vbyte_store(uint8_t *p, uint32_t value);
 
 #endif
